Wrap a REST API with GraphQL
-use yts.mx API
-remove Mutation
-remove mock data
-install axios to fetch movies data from yts.ms API
-modify schema and resolvers
