+++ graphql/schema.graphql
@@ -1,15 +1,19 @@
 type Movie {
   id: Int!
-  name: String!
-  score: Int!
+  title: String!
+  rating: Float!
+  description_intro: String
+  language: String!
+  medium_cover_image: String!
+  genres: [String]
 }
 
 type Query {
-  movies: [Movie]!
+  movies(limit: Int, rating: Float): [Movie]!
   movie(id: Int!): Movie
+  suggestions(id: Int!): [Movie]!
 }
-
-type Mutation {
-  addMovie(name: String!, score: Int!): Movie!
-  deleteMovie(id: Int!): Boolean!
-}
+# type Mutation {
+#   addMovie(name: String!, score: Int!): Movie!
+#   deleteMovie(id: Int!): Boolean!
+# }
